add documentation
The README.md file is the only entry point for the documentation.
Using the Markdown format and the 'ronn' gem, it will be correctly
rendered on the Github project page, converted to roff for the manpage
(optional) and to a static HTML page (for the gh-pages branch).

Note: ronn is located here https://github.com/rtomayko/ronn and can be
installed with:

    gem install ronn