sched: do not setup stdin if it refers to a tty
i3blocks will segfault if it is run in a terminal. This is because stdin
receives an EOF and thus triggers the SIGIO signal (meaning stdin is
ready to be read). The handle_click() function assumes that there's a
json string to read.

This commit do not setup the event I/O on stdin if it refers to a tty.

Fixes #7