make struct click members char buffers
This commit turns the struct click members from unsigned int into char
buffers. A json_parse() function is introduced, which simplifies the
click handling and parsing.

By doing this, we remove the check for a click in the setup_env()
function and thus set the variables BLOCK_{BUTTON,X,Y} to an empty
string when no click happened.

Fixes #9

Reported-by: Imran Sobir <ahmedimransobir@gmail.com>