block: better failure message
This commit adds the exit code (if given) to the short_text,
concatenates the full_text with it and adds the first line of the
child's output (if any) as the error message, if the returned code is
different from 0 or 33. (stderr is still ignored)

In the meantime, rename the failed() function to mark_as_failed(),
switch its arguments and refactor the block_update() function to remove
sprintf/goto calls in favor of return mark_as_failed() (void).