add support for persistent blocklets
This commit adds a new "persist" special interval value. This is meant
for block commands which don't exit and continuously output update (like
daemons). The only con argument is that these block updates are
restricted to a single line per update.

For instance, here's a simple and stupid example:

    [persistent]
    command=while :; do sleep 3; date; done
    interval=persist

This new behavior is handy for monitoring processes such as
`dbus-monitor -m` or `inotifywait -m`.

Refs #85