block: add support for JSON output format
This commit is based on the initial work from @Zopieux, adding support
for JSON output format.

It was rebased on top of the latest changes and has been slightly
changed to rename the "output" property to "format" (since "output"
might be confusing with i3's vocabulary).

A blocklet can now specify its output as JSON like so:

    [date-json]
    command=date +'{"full_text":"%s"}'
    interval=2
    format=json

Refs #85
Closes #87