scripts: (bandwidth) add support for custom IN and OUT labels
The script now can get two command line arguments, -i/--inlabel and
-o/--outlabel to enable customization of the "IN" and "OUT" hardcoded
text.

Note that the hardcoded space is not automatically inserted anymore
before $INLABEL and $OUTLABEL, but to keep the old script behaviour,
default values have been added with $INLABEL="IN " and $OUTLABEL="OUT "
so it behaves exactly the same if no arguments are provided.

Reviewed-by: @kb100

Closes: #129