Promote i3blocks-contrib in README
This commit puts a link to the i3blocks-contrib project directly in the
README. Previously a link was only found in the wiki. This brings users
one click closer to i3blocks-contrib, giving them easier access to many
blocklet examples, as well as promoting use of and contributions to the
sister project.