scripts: (battery) try to guess status by quering ac adapter
The battery script uses `acpi -b` to query to battery status, but the
status might be 'Unknown', especially when the AC adapter has just been
(un)plugged. The script now tries to query the AC adapter status when
that battery status is encountered.

With that modification, it is possible to hook on ACPI events to refresh
the battery status, and make i3blocks more reactive and display useful
information. For example, with `acpid` and the following event files:

  $ cat /etc/acpid/events/battery
  event=ac_adapter
  action=pkill i3blocks -RTMIN+11

and hooking the battery block on signal 11, the battery block will be
refreshed when the AC adapter is (un)plugged.

Closes #205