json: skip array delimiters
The JSON parser does not support parsing object or array at the moment,
since this use case is not necessary for i3bar.

However i3bar shoots clicks in an infinite array, the first line being
an opening bracket. This is not skipped by i3blocks currently, resulting
in an error sent on the first click. To avoid that, skip the '[' and ']'
array delimiters in addition to the '{' and '}' object delimiters.

Fixes #286