block: remove support for label
The "label" property was specific to i3blocks and implemented as an easy
way to prefix the "full_text" property. However it brings some
inconsistencies to the user, such as the calculation of the "min_width".

For this reason, The support for the "label" key is removed.
Below are examples how to properly add a label for a "time" block.

Either you can keep the "label" key and use its value in the command:

    [time]
    label=Time:
    command=echo -n "$label"; date +%T
    interval=30

Or more cleanly with a static block on its left:

    [time-label]
    separator_block_width=0
    separator=false
    full_text=Time:

    [time]
    command=date +%T
    interval=30

Refs #219