click: fix first click handling
The very first line on standard input is the beginning of an infinite
array, thus the first click parsed was empty and processing was stopped.
Hence handling the first click was delayed until the second click.

Now that -EAGAIN is properly propagated up to the click parsing code, we
don't have to break on empty clicks and can then handle them all.

Refs #286