click: handle all properties
Rework the click handling code so that we don't filter properties passed
by i3bar anymore. This adds support for the new "relative_x",
"relative_y", "width", and "height" click properties, which are now
available in the script environment through variables of the same name.

The following block now works as expected:

    [clickme]
    full_text=Click me!
    command=echo button=$button x=$x y=$y relative_x=$relative_x relative_y=$relative_y width=$width height=$height

Closes #302