sched: close stdin on exec
When forking processes, the best practice is to close the standard input
from within the child process and reopen /dev/null or a pipe to provide
a valid file descriptor 0.

Another solution would've been to set the FD_CLOEXEC flag on the
parent's standard input so that the child stdin gets automatically
closed when a process is executed. A valid file descriptor 0 still needs
to be somehow provided though.

Both solutions fix a weird behavior caused by processes like "fish -c
date" messing around with the parent standard input through the
duplicated child stdin and making it blocking again, freezing i3blocks.

This commit implements the second solution for reference even though the
former solution is already implemented. This doesn't hurt to have both
since the child standard input won't be closed again if it already is.

Refs #288