docs: centralize all documentation
This commit moves the README and the manpage source files to a docs/
directory, since Github renders the README from there too.

This allows us to generated the web site pages from the README file and
the manpage as well.

From now on commit every generated Roff and HTML files to ease packaging
and web site deployment.

A pre-commit hook is added to re-generate the documentation if a source
file has staged changes. But the generated changes will be automatically
added to the index only if there are no unstaged source hunks though.
The hook can be installed with:

    ln -s ../../docs/pre-commit $(git rev-parse --git-dir)/hooks/