block: restore deprecated label
Now that user variables are supported, it is preferable that a blocklet
handles its label on its own, even more since this brings inconsistency
with the full_text length calculation.

But at the same time, removing its support right now can bring some
frustration to the user as well as a bad experience. Now that we
store unescaped JSON values again, bring back the support for the
label property and deprecate it for the duration of the next release.

Refs #219