json: skip array values
i3bar sends an array of active modifiers when a click happened.

Currently, i3blocks does not support for array and object values. When
an array is met, the parser returns an error and the event is ignored.

This commit skips over the array without adding it to the map.

Fixes #335
Refs #338