config: fix memory leak
Valgrind complains about the leak below:

	==18102== 179 (8 direct, 171 indirect) bytes in 1 blocks are definitely lost in loss record 20 of 29
	==18102==    at 0x4839B65: calloc (vg_replace_malloc.c:752)
	==18102==    by 0x10E0CF: map_create (map.c:199)
	==18102==    by 0x10CD40: config_set (config.c:75)
	==18102==    by 0x10CD40: config_ini_property_cb (config.c:103)
	==18102==    by 0x10DCEA: line_parse (line.c:68)
	==18102==    by 0x10DD31: line_read (line.c:83)
	==18102==    by 0x10D2B0: ini_read (ini.c:105)
	==18102==    by 0x10CBB1: config_read (config.c:110)
	==18102==    by 0x10CBB1: config_open (config.c:142)
	==18102==    by 0x10CE6A: config_load (config.c:163)
	==18102==    by 0x10AEF5: bar_load (bar.c:342)
	==18102==    by 0x10A493: main (main.c:65)

The stack allocated config structure (which is used locally in
config_load() to parse the configuration file) has a global attribute
that is allocated in config_set() function but never free'd.

The allocation is triggered in config_read() by function ini_read()
through the callbacks config_ini_section_cb() and config_set().

The free should be performed in the same fonction (i.e. config_read()),
but it makes the code more complicated to cover up all error cases and
make sure the allocation is free'd.

The free is performed in config_open() instead that let the code more
trivial.