sched: fix SIGRTMIN and SIGIO on exit
When signals SIGINT or SIGTERM is received, i3blocks ends with SIGRTMIN
or SIGIO return code instead of exiting properly.

This situation happens always with persistent blocks or happens if data
is received on STDIN.

Here is an example that interrupts a single persistent block with
Ctrl+C:

	$ i3blocks -c persist.config
	{"version":1,"click_events":true}
	[[]
	,[{"full_text":""}]
	,[{"full_text":""},{"":"","name":"persist","command":"while :; do echo $name; sleep 1; done","interval":"persist","full_text":"persist"}]
	(...)
	^CReal-time signal 0

Where persit.config is as follow:

	[persist]
	command=while :; do echo $name; sleep 1; done
	interval=persist

Upon the reception of signals SIGINT or SIGTERM, the scheduler exits
from the main loop causes signals to be unblocked.

But persistents blocks raise signal to i3blocks because on events on
their STDOUT that lead i3blocks to exit with SIGRTMIN.

The same situation happens if data is received from STDIN that leads
i3blocks to exit with SIGIO.

To prevent from such situations, the signals that are set to file
descriptors must be unset before the signal are unblocked.

First, this commit improves the existing function sys_async(). When the
signal given in parameter is 0, it removes the ASYNC and NONBLOCK flags
and the signal set to the file descriptor.

Then, the commit fixes the scheduler using the improvement above. The
signals that are set to file descriptors are unset before before to exit
the loop. That makes sure i3blocks will not be signaled with SIGRTMIN
and SIGIO.