json: unescape at parsing
Rework the JSON implementation so that the parsing code can
optionally write the unescape value to a given buffer.

Add basic support for decoding a single unicode codepoint.

This makes the json_unescape and json_pair functions obsolete, as
well as the json data structure.

Also use the new parsing code to validate arbitrary JSON values.

Refs #345