block, json, log: fix different signedness comparisons
This fixes 2 comparisons between different signedness integers.

The first one is the signal type defined in block structure. In C,
signal numbers are defined as int. This commit fixes the definition in
block.h.

The second one is a comparison between and int and an enum. The
comparison requires the enum to be caseted in int.

The third one is the comparison between the result of snprintf()(ie.
int) and a sizeof variable (ie. size_t). This comparison requires the
int to be casted in unsigned, since the if statement ensures that the
value is not negative. This commit casts the int variable in size_t when
it is compared to a size_t.