readme: repeat interval does not support clicks
Mention in the README (and thus the website) that the repeat interval
does not support the clicks, since such scripts are unlikely to read
their stdin and even if they do, it will still be racy because of the
reap and respawn time. Killing and restarting the commands isn't an
option either, since they may have an heavy initialization time and
that won't avoid the raciness anyway.

Refs #382