add a minimal system-wide configuration file
In order not to have an empty bar on a fresh installation of i3blocks,
include a basic system-wide configuration to fallback to. It can be
overriden by package maintainers for sure.

Note that it must not bring any dependency to the i3blocks package,
that is not what it is meant for.

Refs #230