Accept EOF as well as newline to complete line read
Currently a line isn't read if any error is reached, including EAGAIN,
which indicates you've reached the end of a file descriptor.

This commit allows EAGAIN to terminate a line if anything else has been
read from that fd. If nothing else has been read, it's either
non-existent or empty, which makes sense to error instead.