block: Keep RTMIN through RTMAX signals blocked in child
When a child process is forked for a block, it has the signals RTMIN
through RTMAX blocked by default, but block_child_sig() unblocks all
signals including these.

Unblocking these signals is troublesome given that:
- The easiest way to signal i3blocks for an update is through pkill (as
  shown at https://vivien.github.io/i3blocks/#_signal ). But pkill sends
  a signal to all processes with that name, which includes the child
  fork.
- Since those signals are unblocked, but not handled, in the child, when
  one of them is sent to i3blocks, and while it is being handled in
  the child process, another one is sent, the child process will also
  receive the signal and therefore crash since it is unhandled.

If we assume that the RTMIN through RTMAX signals are uncommon enough
that they won't be used in the child process, the solution is as easy as
keeping these signals blocked for the child process.