bar: do not create an empty block 'foo'
The current code was creating an empty block to simplify the addition
of new blocks from the config callback. This resulted in an empty
'foo' block being created which is confusing to others.

Do not create this empty block anymore since the config callback now
assigns bar->blocks if there is no block created yet.

Refs #411