line: return -EINVAL on incomplete lines
When the line reading code reaches the end of file or pipe, -EAGAIN
is returned and eventual characters are silently ignored.

This can be confusing for a user wrongly writing lines not ending
with a newline character, or simply if an error occured like a buffer
flushed too soon.

In order to distinguish this, error out the incomplete line and
return -EINVAL.