bar: block all signals
Instead of starting from an empty set of signal and adding in it
only the signals we care about, start from the full set of signals
are mask them all.

Add a new parameter to the sigprocmask wrapper in order to store the
original set of signals in bar->sigset, so that we can restore it
on teardown.