feat: allow writing custom text to stdin on button press
Rather than expecting external programs to handle 1, 2, and 3 on
their standard input, this allows the user to configure the text
sent themselves.

This is particularly relevant for my tool, painted, which serves
as a notification daemon and takes commands to manage the notif
queue over stdin. With this patch, a config like the following
works just as one would expect:

	[notifs]
	command=painted
	interval=persist

	btn_left_cmd=clear
	btn_right_cmd=expand

Note this only applies to bars with the persist interval, as one
would expect.